requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
Jinja2==2.11.1
jinja2-cli==0.7.0
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.2
psycopg2-binary==2.8.5
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
natsort==7.0.1
tabulate==0.8.6
packaging==20.4
sqlparse==0.3.1
asyncpg==0.21.0
python-dateutil==2.8.0
